pyproject.toml
[project]
name = "py-nfse"
version = "0.2.0"
description = "Agente para download automatico de NFSe com certificado digital"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "playwright>=1.40.0",
    "pywin32>=306; sys_platform == 'win32'",
    "pyautogui>=0.9.54",
    "pywinauto>=0.6.8; sys_platform == 'win32'",
    "oracledb>=3.4.2",
    "pandas>=3.0.0",
    "openpyxl>=3.1.5",
]
